Replace setup.py with equivalent poetry config
The manifest file has been replaced with `include` config. To check this
was sensible, I compared the sdist on pypi against the sdist generated
by `poetry build`. I adjusted the TOML config until the newer sdist
matched the existing one.

We don't commit the lockfile, because this is a library.